feat(main): use f64 and cast i32 into it

+++ src/main.rs
@@ -1,9 +1,9 @@
 fn main() {
-    let mut sum = 0;
+    let mut sum = 0.0;
     for i in 0..5 {
         let even_odd = if i & 1 == 0 { "even" } else { "odd" };
         println!("{} {}", even_odd, i);
-        sum += i;
+        sum += i as f64;
     }
     println!("Sum: {}", sum);
 }
